+++ RNSentryAndroidTester/app/src/test/java/com/swmansion/rnscreens/ScreenStackFragment.kt
@@ -1,0 +1,7 @@
+package com.swmansion.rnscreens
+
+import androidx.fragment.app.Fragment
+
+class ScreenStackFragment(contentLayoutId: Int) : Fragment(contentLayoutId) {
+
+}