+++ android/src/main/java/io/sentry/react/RNSentryPackage.java
@@ -1,8 +1,11 @@
 package io.sentry.react;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import com.facebook.react.bridge.NativeModule;
@@ -10,6 +13,7 @@ import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.module.model.ReactModuleInfo;
 import com.facebook.react.module.model.ReactModuleInfoProvider;
 import com.facebook.react.TurboReactPackage;
+import com.facebook.react.uimanager.ViewManager;
 
 public class RNSentryPackage extends TurboReactPackage {
 
@@ -41,6 +45,15 @@ public class RNSentryPackage extends TurboReactPackage {
       ));
       return moduleInfos;
     };
+  }
+
+  @NonNull
+  @Override
+  public List<ViewManager> createViewManagers(
+          ReactApplicationContext reactContext) {
+    return Arrays.asList(
+          new RNSentryOnDrawReporterManager(reactContext)
+    );
   }
 
 }