+++ samples/react-native/android/app/src/main/java/com/samplenewarchitecture/NativePlatformSampleModule.kt
@@ -1,0 +1,17 @@
+package com.samplenewarchitecture
+
+import com.facebook.fbreact.specs.NativePlatformSampleModuleSpec
+import com.facebook.react.bridge.ReactApplicationContext
+
+class NativePlatformSampleModule(reactContext: ReactApplicationContext) : NativePlatformSampleModuleSpec(reactContext) {
+
+    override fun getName() = NAME
+
+    override fun crashOrString(): String {
+        throw RuntimeException("JVM Crash in NativePlatformSampleModule.crashOrString()")
+    }
+
+    companion object {
+        const val NAME = "NativePlatformSampleModule"
+    }
+}