+++ samples/react-native/src/Screens/TrackerScreen.tsx
@@ -6,6 +6,7 @@ import {
   Text,
   ActivityIndicator,
 } from 'react-native';
+import delay from 'delay';
 
 import * as Sentry from '@sentry/react-native';
 
@@ -17,16 +18,19 @@ import * as Sentry from '@sentry/react-native';
  * to the fetch call and track the time it takes for Promise to resolve.
  */
 const TrackerScreen = () => {
+  const [state, setState] = React.useState<'loading' | 'loaded' | 'error'>(
+    'loading',
+  );
   const [cases, setCases] = React.useState<{
     TotalConfirmed: number;
     TotalDeaths: number;
     TotalRecovered: number;
   } | null>(null);
 
-  const loadData = () => {
+  const loadData = async () => {
     setCases(null);
 
-    fetch('https://api.covid19api.com/summary', {
+    const maybeData = fetch('https://api.covid19api.com/summary', {
       method: 'GET',
       headers: {
         Accept: 'application/json',
@@ -37,6 +41,15 @@ const TrackerScreen = () => {
       .then(json => {
         setCases(json.Global);
       });
+
+    try {
+      await Promise.allSettled([maybeData, delay(2_000)]);
+      await maybeData;
+      setState('loaded');
+    } catch (e) {
+      Sentry.captureException(e);
+      setState('error');
+    }
   };
 
   const onRefreshButtonPress = () => {
@@ -50,8 +63,18 @@ const TrackerScreen = () => {
     loadData();
   }, []);
 
+  const statusText =
+    (state === 'loading' && 'Loading...') ||
+    (state === 'error' && 'Error') ||
+    (state === 'loaded' && 'Loaded') ||
+    'Unknown';
+  const shouldRecordFullDisplay = state === 'loaded' || state === 'error';
+  console.log('shouldRecordFullDisplay', shouldRecordFullDisplay);
+  console.log('statusText', statusText);
+
   return (
     <View style={styles.screen}>
+      <Sentry.TimeToInitialDisplay record />
       <View style={styles.titleContainer}>
         <Text style={styles.title}>Global COVID19 Cases</Text>
       </View>
@@ -78,11 +101,14 @@ const TrackerScreen = () => {
           <ActivityIndicator size="small" color="#F6F6F8" />
         )}
       </View>
-      <Button
-        sentry-label="refresh"
-        title="Refresh"
-        onPress={onRefreshButtonPress}
-      />
+      <Sentry.TimeToFullDisplay record={shouldRecordFullDisplay}>
+        <Button
+          sentry-label="refresh"
+          title="Refresh"
+          onPress={onRefreshButtonPress}
+        />
+        <Text>{statusText}</Text>
+      </Sentry.TimeToFullDisplay>
     </View>
   );
 };