+++ src/js/client.ts
@@ -6,10 +6,8 @@ import type {
   Envelope,
   Event,
   EventHint,
-  Exception,
   Outcome,
   SeverityLevel,
-  Thread,
   UserFeedback,
 } from '@sentry/types';
 import { dateTimestampInSeconds, logger, SentryError } from '@sentry/utils';
@@ -58,22 +56,7 @@ export class ReactNativeClient extends BaseClient<ReactNativeClientOptions> {
    * @inheritDoc
    */
   public eventFromMessage(message: string, level?: SeverityLevel, hint?: EventHint): PromiseLike<Event> {
-    return eventFromMessage(this._options.stackParser, message, level, hint, this._options.attachStacktrace).then(
-      (event: Event) => {
-        // TMP! Remove this function once JS SDK uses threads for messages
-        if (!event.exception?.values || event.exception.values.length <= 0) {
-          return event;
-        }
-        const values = event.exception.values.map(
-          (exception: Exception): Thread => ({
-            stacktrace: exception.stacktrace,
-          }),
-        );
-        (event as { threads?: { values: Thread[] } }).threads = { values };
-        delete event.exception;
-        return event;
-      },
-    );
+    return eventFromMessage(this._options.stackParser, message, level, hint, this._options.attachStacktrace);
   }
 
   /**
@@ -134,7 +117,7 @@ export class ReactNativeClient extends BaseClient<ReactNativeClientOptions> {
     }
 
     let shouldClearOutcomesBuffer = true;
-    if (this._transport && this._dsn) {
+    if (this._isEnabled() && this._transport && this._dsn) {
       this.emit('beforeEnvelope', envelope);
 
       this._transport.send(envelope).then(null, reason => {
@@ -185,7 +168,7 @@ export class ReactNativeClient extends BaseClient<ReactNativeClientOptions> {
     if (__DEV__ && this._options.enableNativeNagger) {
       Alert.alert(
         'Sentry',
-        'Warning, could not connect to Sentry native SDK.\nIf you do not want to use the native component please pass `enableNative: false` in the options.\nVisit: https://docs.sentry.io/platforms/react-native/#linking for more details.',
+        'Warning, could not connect to Sentry native SDK.\nIf you do not want to use the native component please pass `enableNative: false` in the options.\nVisit: https://docs.sentry.io/platforms/react-native/ for more details.',
       );
     }
   }