+++ src/js/integrations/debugsymbolicator.ts
@@ -60,7 +60,6 @@ export class DebugSymbolicator implements Integration {
         symbolicatedFrames && this._replaceExceptionFramesInEvent(event, symbolicatedFrames);
       } else if (event.threads) {
         // RN JS doesn't have threads
-        // syntheticException is used for Sentry.captureMessage() threads
         symbolicatedFrames && this._replaceThreadFramesInEvent(event, symbolicatedFrames);
       }
     }
@@ -219,7 +218,17 @@ export class DebugSymbolicator implements Integration {
           if (xhr.status !== 200) {
             resolve(null);
           }
-          resolve(xhr.responseText);
+          const response = xhr.responseText;
+          if (
+            typeof response !== 'string' ||
+            // Expo Dev Server responses with status 200 and config JSON
+            // when web support not enabled and requested file not found
+            response.startsWith('{')
+          ) {
+            resolve(null);
+          }
+
+          resolve(response);
         }
       };
       xhr.onerror = (): void => {