+++ src/js/wrapper.ts
@@ -10,7 +10,7 @@ import type {
   User,
 } from '@sentry/types';
 import { logger, normalize, SentryError } from '@sentry/utils';
-import { NativeModules, Platform, TurboModuleRegistry } from 'react-native';
+import { NativeModules, Platform } from 'react-native';
 
 import { isHardCrash } from './misc';
 import type {
@@ -27,11 +27,19 @@ import type * as Hermes from './profiling/hermes';
 import type { NativeAndroidProfileEvent, NativeProfileEvent } from './profiling/nativeTypes';
 import type { RequiredKeysUser } from './user';
 import { isTurboModuleEnabled } from './utils/environment';
+import { ReactNativeLibraries } from './utils/rnlibraries';
 import { base64StringFromByteArray, utf8ToBytes } from './vendor';
 
-const RNSentry: Spec | undefined = isTurboModuleEnabled()
-  ? TurboModuleRegistry.get<Spec>('RNSentry')
-  : NativeModules.RNSentry;
+/**
+ * Returns the RNSentry module. Dynamically resolves if NativeModule or TurboModule is used.
+ */
+export function getRNSentryModule(): Spec | undefined {
+  return isTurboModuleEnabled()
+    ? ReactNativeLibraries.TurboModuleRegistry && ReactNativeLibraries.TurboModuleRegistry.get<Spec>('RNSentry')
+    : NativeModules.RNSentry;
+}
+
+const RNSentry: Spec | undefined = getRNSentryModule();
 
 export interface Screenshot {
   data: Uint8Array;
@@ -95,6 +103,7 @@ interface SentryNativeWrapper {
    * Fetches native stack frames and debug images for the instructions addresses.
    */
   fetchNativeStackFramesBy(instructionsAddr: number[]): NativeStackFrames | null;
+  initNativeReactNavigationNewFrameTracking(): Promise<void>;
 }
 
 const EOL = utf8ToBytes('\n');
@@ -586,6 +595,17 @@ export const NATIVE: SentryNativeWrapper = {
     }
 
     return RNSentry.fetchNativeStackFramesBy(instructionsAddr) || null;
+  },
+
+  async initNativeReactNavigationNewFrameTracking(): Promise<void> {
+    if (!this.enableNative) {
+      return;
+    }
+    if (!this._isModuleLoaded(RNSentry)) {
+      return;
+    }
+
+    return RNSentry.initNativeReactNavigationNewFrameTracking();
   },
 
   /**